chatbot/openai/requirements.txt
fastapi
uvicorn[standard]
redis
requests
langchain-openai
langchain-community
langchain-core
tiktoken
pgvector
python-dotenv
psycopg2-binary
